Rename `classCustomHandler.py` -> `handlers.py`.
