+++ logging.ini
@@ -28,12 +28,12 @@ formatter=formatter_debug
 args=(sys.stdout,)
 
 [handler_customHandler_debug]
-class=src.logging.classCustomHandler.CustomTimedRotatingFileHandler
+class=src.logging.handlers.CustomTimedRotatingFileHandler
 formatter=formatter_debug
 args = ("Debug-","D", 7)
 
 [handler_customHandler_info]
-class=src.logging.classCustomHandler.CustomTimedRotatingFileHandler
+class=src.logging.handlers.CustomTimedRotatingFileHandler
 formatter=formatter_info
 args = ("Debug-","D", 7)
